+++ package.json
@@ -19,22 +19,21 @@
   "license": "MIT",
   "bin": "dist/main.js",
   "scripts": {
-    "autoclean:all": "yarn autoclean --force && cd frontend && yarn autoclean --force && cd ../tray && yarn autoclean --force",
-    "build": "yarn build:frontend && yarn build:app && yarn build:tray",
+    "autoclean:all": "yarn autoclean --force && cd frontend && yarn autoclean --force && yarn autoclean --force",
+    "build": "yarn build:frontend && yarn build:app",
     "build:app": "rimraf dist && nest build",
     "build:frontend": "cd frontend && yarn build",
-    "build:tray": "cd tray && yarn build",
     "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
-    "install:all": "yarn install && cd frontend && yarn install && cd ../tray && yarn install",
-    "install:all:ci": "yarn install --pure-lockfile && cd frontend && yarn install --pure-lockfile && cd ../tray && yarn install --pure-lockfile",
-    "upgrade:all": "yarn upgrade --latest --exact && cd frontend && yarn upgrade --latest --exact && cd ../tray && yarn upgrade --latest --exact",
-    "upgrade-interactive:all": "yarn upgrade-interactive --latest --exact && cd frontend && yarn upgrade-interactive --latest --exact && cd ../tray && yarn upgrade-interactive --latest --exact",
+    "install:all": "yarn install && cd frontend && yarn install",
+    "install:all:ci": "yarn install --pure-lockfile && cd frontend && yarn install --pure-lockfile",
+    "upgrade:all": "yarn upgrade --latest --exact && cd frontend && yarn upgrade --latest --exact",
+    "upgrade-interactive:all": "yarn upgrade-interactive --latest --exact && cd frontend && yarn upgrade-interactive --latest --exact",
     "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
     "package": "rimraf out && yarn build && node scripts/package.js",
     "package:app": "rimraf out && yarn build:app && cross-env APP_ONLY=true node scripts/package.js",
-    "start": "concurrently -k \"yarn start:app\" \"yarn start:frontend\" \"yarn start:tray\"",
-    "start:dev": "concurrently -k \"yarn start:app:dev\" \"yarn start:frontend:dev\" \"yarn start:tray\"",
-    "start:prod": "concurrently -k \"yarn start:app:prod\" \"yarn start:tray:prod\"",
+    "start": "concurrently -k \"yarn start:app\" \"yarn start:frontend\"",
+    "start:dev": "concurrently -k \"yarn start:app:dev\" \"yarn start:frontend:dev\"",
+    "start:prod": "cross-env SB_USE_CWD=true node dist/",
     "start:app": "cross-env NODE_ENV=development nest start",
     "start:app:cli": "cross-env NODE_ENV=development SB_CLI=true SB_TRAY=false nest start",
     "start:app:cli:dev": "cross-env NODE_ENV=development SB_CLI=true SB_TRAY=false nest start --watch",
@@ -45,8 +44,6 @@
     "start:cli:dev": "concurrently -k \"yarn start:app:cli:dev\"",
     "start:frontend": "cd frontend && yarn start",
     "start:frontend:dev": "cd frontend && yarn start:dev",
-    "start:tray": "cd tray && cross-env NODE_ENV=development yarn start",
-    "start:tray:prod": "cd tray && cross-env SB_USE_CWD=true yarn start:prod",
     "test": "jest",
     "test:cov": "jest --coverage",
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
@@ -84,6 +81,7 @@
     "swagger-ui-express": "4.1.6",
     "system-bridge-windows-sensors": "1.2.1",
     "systeminformation": "5.9.3",
+    "systray2": "2.1.3",
     "toad-scheduler": "1.5.0",
     "typeorm": "0.2.37",
     "uuid": "8.3.2",
