+++ src/components/common.ts
@@ -90,8 +90,7 @@ export async function createSetting(
 }
 
 export async function getUpdates(
-  logger: Logger,
-  tray = false
+  logger: Logger
 ): Promise<ApplicationUpdate | undefined> {
   try {
     const response = await axios.get<{
@@ -105,7 +104,7 @@ export async function getUpdates(
       response.data?.prerelease === false
     ) {
       const versionNew = semver.clean(response.data.tag_name);
-      const versionCurrent = getVersion(logger, tray);
+      const versionCurrent = getVersion(logger);
       const available = semver.lt(versionCurrent, versionNew);
       return {
         available,
@@ -122,7 +121,7 @@ export async function getUpdates(
   return undefined;
 }
 
-export function getVersion(logger: Logger, tray = false): string {
+export function getVersion(logger: Logger): string {
   try {
     const json = JSON.parse(
       readFileSync(
@@ -136,7 +135,7 @@ export function getVersion(logger: Logger, tray = false): string {
                   ? process.execPath.lastIndexOf("\\")
                   : process.execPath.lastIndexOf("/")
               ),
-          tray ? "../package.json" : "package.json"
+          "package.json"
         ),
         {
           encoding: "utf8",
