--- tray/package.json
@@ -1,48 +1,0 @@
-{
-  "name": "system-bridge-tray",
-  "version": "2.6.0",
-  "description": "A bridge for your systems",
-  "private": true,
-  "author": {
-    "name": "Aidan Timson <Timmo>",
-    "email": "contact@timmo.xyz",
-    "url": "https://system-bridge.timmo.dev"
-  },
-  "keywords": [
-    "system",
-    "bridge"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/timmo001/system-bridge"
-  },
-  "license": "MIT",
-  "bin": "dist/tray.js",
-  "scripts": {
-    "build": "rimraf dist && tsc",
-    "start": "yarn build && node dist/tray.js",
-    "start:prod": "node dist/tray.js"
-  },
-  "dependencies": {
-    "better-sqlite3": "7.4.3",
-    "open": "8.2.1",
-    "query-string": "7.0.1",
-    "rimraf": "3.0.2",
-    "systray2": "2.1.2",
-    "typeorm": "0.2.37",
-    "ws": "7.5.4"
-  },
-  "devDependencies": {
-    "pkg": "5.3.1",
-    "typescript": "4.4.2"
-  },
-  "pkg": {
-    "assets": [
-      "node_modules/**/*"
-    ],
-    "scripts": "dist/tray.js",
-    "targets": [
-      "host"
-    ]
-  }
-}
