--- tray/src/components/entities/applicationUpdate.entity.ts
@@ -1,6 +1,0 @@
-export interface ApplicationUpdate {
-  available: boolean;
-  newer: boolean;
-  url: string;
-  version: { current: string; new: string };
-}
