--- tray/src/components/entities/event.entity.ts
@@ -1,4 +1,0 @@
-export interface Event {
-  name: string;
-  data?: any;
-}
