--- tray/src/components/entities/setting.entity.ts
@@ -1,10 +1,0 @@
-import { Entity, Column, PrimaryColumn } from "typeorm";
-
-@Entity()
-export class Setting {
-  @PrimaryColumn({ type: "varchar" })
-  key = "";
-
-  @Column({ type: "varchar" })
-  value = "";
-}
