--- tray/tsconfig.build.json
@@ -1,4 +1,0 @@
-{
-  "exclude": ["node_modules", "dist"],
-  "extends": "./tsconfig.json"
-}
