--- tray/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "commonjs",
-    "allowSyntheticDefaultImports": true,
-    "declaration": true,
-    "emitDecoratorMetadata": true,
-    "esModuleInterop": true,
-    "experimentalDecorators": true,
-    "removeComments": true,
-    "sourceMap": true,
-    "target": "ES2021",
-    "outDir": "./dist"
-  }
-}
